Move custom roles to the separate folder
