+++ src/ggrc-client/js/models/custom-roles/index.js
@@ -1,0 +1,7 @@
+/*
+  Copyright (C) 2018 Google Inc.
+  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+*/
+
+import './access-control-role';
+import './roleable';
